+++ SWP391_WebXemPhimTrucTuyen/src/main/java/controller/UserController.java
@@ -12,7 +12,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import entity.User;
+import model.User;
 import jakarta.servlet.http.HttpSession;
 
 /**
