+++ SWP391_WebXemPhimTrucTuyen/src/main/java/util/DBContext.java
@@ -1,0 +1,62 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package util;
+
+
+import java.sql.Connection;
+import java.sql.DatabaseMetaData;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+
+/**
+ *
+ * @author Chau Tan Cuong - CE190026
+ */
+public class DBContext {
+
+    private Connection conn;
+    private final String DB_URL = "jdbc:sqlserver://127.0.0.1:1433;databaseName=NetFlopDB;encrypt=false";
+    private final String DB_USER = "sa";
+    private final String DB_PWD = "7355608";
+
+    public DBContext() {
+        try {
+            Class.forName("com.microsoft.sqlserver.jdbc.SQLServerDriver");
+            this.conn = (Connection) DriverManager.getConnection(DB_URL, DB_USER, DB_PWD);
+        } catch (ClassNotFoundException | SQLException ex) {
+            Logger.getLogger(DBContext.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
+
+    public Connection getConnection() {
+        return conn;
+    }
+
+    public static void closeConnection(Connection c) {
+        try {
+            if (c != null) {
+                c.close();
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
+
+    public static void printInfo(Connection c) throws SQLException {
+        try {
+            if (c != null) {
+                DatabaseMetaData mtdt = c.getMetaData();
+                System.out.println(mtdt.getDatabaseProductName());
+                System.out.println(mtdt.getDatabaseProductVersion());
+
+            }
+        } catch (SQLException e) {
+            e.printStackTrace();
+        }
+    }
+}
